Fixed file locationq
- Fixed location of locales from `Prefabs` to `Resources`
- Fixed location of menu script from `Prefabs/Player` to `Scripts`
