+++ Assets/Scripts/PlayerSelectionMenu/PlayerSelection.cs
@@ -1,0 +1,42 @@
+using UnityEngine;
+using UnityEngine.SceneManagement;
+using UnityEngine.UI;
+
+//useful for the buttons
+namespace Prefabs.UI
+{
+    public class PlayerSelection : MonoBehaviour
+    {
+        private static int currentPlayer = 0;
+        [SerializeField] private Button previousPlayerButton;
+        [SerializeField] private Button nextPlayerButton;
+        [SerializeField] public GameObject prefabPlayer;
+
+        private void Awake()
+        {
+            for (int i = 0; i < transform.childCount; i++)
+            {
+                transform.GetChild(i).gameObject.SetActive(i==currentPlayer);
+            }
+        }
+    
+        public void ChangeMyPlayer(int change) //for my buttons < and >
+        {
+            transform.GetChild(currentPlayer).gameObject.SetActive(false);
+            currentPlayer += change + transform.childCount; //will set the index to + change 
+            currentPlayer %= transform.childCount;
+            transform.GetChild(currentPlayer).gameObject.SetActive(true);
+        }
+
+        public void ChangeScene(string sceneName)
+        {
+            Transform child = prefabPlayer.transform.GetChild(0);
+            Debug.Log(currentPlayer);
+            for (int i = 1; i < child.childCount; i++)
+            {
+                child.GetChild(i).gameObject.SetActive(currentPlayer + 1 == i);
+            }
+            SceneManager.LoadScene(sceneName);
+        }
+    }
+}
